requirements.txt
contourpy==1.3.0
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.58.0
importlib_resources==6.5.2
jieba==0.42.1
kiwisolver==1.4.7
matplotlib==3.9.4
networkx==3.2.1
numpy==2.0.2
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pyparsing==3.2.3
pypinyin==0.54.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
tzdata==2025.2
wordcloud==1.9.4
zipp==3.21.0
